Rework the signature of WalkFn to allow setting new values
Before this commit, you could change the nodes during walk, assuming they were pointers.
But it was not possible to replace a node with a new one (unless they were stored as *interface{}, but that doesn's sound practical).

You could do a new `Insert`, but that be data racy if you e.g. partitioned the tree and walked and modified it concurrently.