Fix potential data race in Walk
The construct in #3 was added to allow setting values while walking but also allow other goroutines to walk the tree concurrently.

But that premiss was flawed, as any walk would read the leaf values, hence causing a potential data race.

This changes makes the Walk take an interface with an option al pre Check method that takes only the key, and a Handle method that takes both key and value.
This allows the caller to skip reading the value if not needed.